+++ test/integration/011_invalid_model_tests/test_invalid_models.py
@@ -1,4 +1,5 @@
 from test.integration.base import DBTIntegrationTest, use_profile
+import os
 
 
 class TestInvalidDisabledModels(DBTIntegrationTest):
@@ -45,3 +46,35 @@ class TestInvalidModelReference(DBTIntegrationTest):
             self.run_dbt()
 
         self.assertIn('which was not found', str(exc.exception))
+
+
+class TestInvalidMacroCall(DBTIntegrationTest):
+    @property
+    def schema(self):
+        return "invalid_models_011"
+
+    @property
+    def models(self):
+        return "models-4"
+
+    @staticmethod
+    def dir(path):
+        return path.lstrip("/")
+
+    @property
+    def project_config(self):
+        return {
+            'macro-paths': [self.dir('bad-macros')],
+        }
+
+    @use_profile('postgres')
+    def test_postgres_call_invalid(self):
+        with self.assertRaises(Exception) as exc:
+            self.run_dbt(['compile'])
+
+
+        macro_path = os.path.join('bad-macros', 'macros.sql')
+        model_path = os.path.join('models-4', 'bad_macro.sql')
+
+        self.assertIn(f'> in macro some_macro ({macro_path})', str(exc.exception))
+        self.assertIn(f'> called by model bad_macro ({model_path})', str(exc.exception))